Add woocommerce/components tree shaking support
