+++ packages/components/package.json
@@ -20,6 +20,10 @@
 	"main": "build/index.js",
 	"module": "build-module/index.js",
 	"react-native": "src/index",
+	"sideEffects": [
+		"build-style/**",
+		"src/**/*.scss"
+	],
 	"dependencies": {
 		"@woocommerce/csv-export": "file:../csv-export",
 		"@woocommerce/currency": "file:../currency",
